Refactor: move utils models to utils/models package
