+++ app/src/main/java/fr/free/nrw/commons/utils/models/NetworkConnectionType.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.utils
+package fr.free.nrw.commons.utils.models
 
 enum class NetworkConnectionType {
     WIFI, TWO_G, THREE_G, FOUR_G, UNKNOWN
